Fix: data display music
Fixed the display of data from the DB  on both artiste and music + added btn for the modal url display.
(WIP):still trying to link both artist and music + tried to add style of music to see how it differs from the other two but so far didn't managed to make it work.